Crate static_analyser

A parser and static analysis library for exploring Delphi code.

This is written using a Literate Programming style, so you may find it easier to inspect the rendered version instead.

Modules

analysis
Static analysis passes (i.e. the whole point).

codemap
A mapping from arbitrary locations and sections of source code to their contents.

errors
Types and traits used for internal errors.

lex
Module for performing lexical analysis on source code.

lowering
Typechecking and basic control flow graph generation.

parse
Parse a stream of Tokens into an Abstract Syntax Tree we can use for the later steps.

Macros

tok
Shorthand macro for generating a token from anything which can be converted into a TokenKind, or any of the TokenKind variants.

Structs

Driver
The driver is in charge of orchestrating the whole analysis process and making sure all the bits and pieces integrate nicely.